+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # dartlang plugin changelog
 
+## unreleased
+- adjusted the order of code completions
+
 ## 0.6.11
 - traverse further down the directory tree looking for Dart projects
 - when looking for Dart projects, ignore any directories that contain a